Split typechecker behavior-ref metadata helpers
